kafka: switch to report full topic configs
This is the main change of the PR. It changes the CreateTopics handler
to use the shared code of DescribeConfigs to generate the configs
returned to the client.

We can share this code because the DescribeConfigs object is a more
general version of the configs returned in the CreateTopics endpoint, so
we can just derive the latter from the former.

The reason for changing the CreateTopics handler this way is that
currently the CreateTopics handler has a bug where it only ever returns
the topic-specific override configs of the topic in the response,
whereas it should return all the configs of the topic (incl both the
topic-specific override configs and the global broker-level-set topic
config defaults). This is the behavour that Apache Kafka exhibits and
the behaviour that KIP 525 prescribes.

Ref:
 * https://cwiki.apache.org/confluence/display/KAFKA/KIP-525+-+Return+topic+metadata+and+configs+in+CreateTopics+response