dt/kafka: refactor fetch quota behaviour test
This test is testing the fetch throttling behaviour, specifically the
fact that a consumer can consume above the configured limit once and any
throttling delays will only apply on the next fetch request.

We want to express the max fetch bytes in terms of data being
consumed, not in terms of the response size (the quotas apply based on
the response size). Therefore, to ensure that we consume enough to go
above the limit, but do so in a single fetch request, we change
the `max_partition_fetch_bytes` of the consumer.

Now, this lets the consumer fetch 11/10=1.1x the data, so we need to
ensure that we have enough data produced both for this fetch and the
second fetch, so we also increase the produced amount of data (to 2x for
simplicity).