kafka/quota_mgr: sharded last_seen_ms
Here, we start to use the shared_value to keep track of when we last saw
a client.id in the client quotas map. The sharded_value is efficient
because all the updates on the request path are done to a local copy
owned by that shard, whereas inside the do_gc() method we can gather
which client ids to expire concurrently to request processing.