k/alter: fix default remove
The deprecated AlterConfigs API differs from the new
IncrementalAlterConfigs API in that it is supposed to clear all configs
not specified in the request. The comment in the function even states
this requirement, but the behaviour was incorrect, as it was trying to
`set` these fields (to the default value of the property's type T) when
it should be `remove`'ing these properties to ensure that they are not
specified at the topic-level and fall back to their cluster-level
defaults.