schema_registry/json: collect_bundled_schema
this recursive function crawl the schema to gather all bundled schemas.
a bundled schema is defined as a schema with "$id" defined.
the recursion is on the "$defs"/"definitions" members of a schema, since
the schema validation ensures that children of these properties are
valid schemas.

some small notes:
- it's possible that bundled schemas use a different
dialect than the root object, so we collect this
- the bundled schemas could be nested
- we can't error in this function since it's called at parse time and
  it's not assured that a bundled schema is actually accessed