cluster: add `topic_table::get_name_by_id(...)`
Implement an efficient lookup for finding a topic by its topic id. This
will be used by kafka handlers to translate the topic id received on a
request to its topic name.

The optional topic id is always checked for nullopt-ness, since the code
needs to work during the migration period when not all topics have topic
ids assigned.