schema_registry: align readwrite mode id behaviour
Removes support for posting schemas with explicit IDs in readwrite mode
to match reference implementation behavior.

New validation logic is added to reject explicit IDs unless in import
mode.

The removed tests were specifically testing the now-invalid behavior of
posting schemas with explicit IDs in readwrite mode.

A new ducktape test is added to demonstrate the expected behaviour of
READWRITE mode when the id is specified while trying to register a
schema.