k/server: keep connection in list while aborting
Extend the time during which a connection is kept on the boost intrusive
list stored in kafka::server. This is to allow including these
connections in the ListKafkaConnections response.

Revoking credentials has been the only other usage of this
`_connections` list, and this now needs a check to skip aborting
connections.