admin: add cluster-wide aggr for ListKafkaConns
Support cluster-wide aggregation of the dataset in the
ListKafkaConnections endpoint. When the `node_id` field in the request
is set to -2, the broker collects the responses from all brokers and
returns a "cluster-wide" view of kafka connections.

Crucially, this cluster-wide view respects the `order_by` field of the
request, which makes it much simpler to build a cluster-wide top-like
utility in rpk.