admin: ListKafkaConnections as ClusterService rpc
Move the endpoint from the BrokerService to the ClusterService, as we
intend to make this endpoint return cluster-wide data.

As part of this, the node_id field on the request is made optional. If
it is not specified, the data returned is always cluster-wide.

The is purely a code-shifting commit. The collection of cluster-wide
data is implemented in the next commit.