admin: ListKafkaConnections as ClusterService rpc
Move the endpoint from the BrokerService to the ClusterService, as we
intend to make this endpoint return cluster-wide data to simplify rpk's
logic of surfacing a cluster-wide view of "top" kafka connections.

As part of this, the node_id field on the request is removed. The
endpoint is always going to return cluster-wide data, and internal
gathering is going to be implemented using `ctx.is_proxied()` in the
next comimt.

The is purely a code-shifting commit. The collection of cluster-wide
data is implemented in the next commit.